+++ seedfarmer/services/sessionmanager.py
@@ -9,6 +9,8 @@ import boto3
 import botocore
 from boto3 import Session
 
+import seedfarmer
+
 _logger: logging.Logger = logging.getLogger(__name__)
 
 
@@ -168,7 +170,7 @@ class SessionManager(ISessionManager, metaclass=SingletonMeta):
             retries={"max_attempts": 5},
             connect_timeout=10,
             max_pool_connections=10,
-            # user_agent_extra=f"seedfarmer/{seedfarmer.__version__}",
+            user_agent_extra=f"seedfarmer/{seedfarmer.__version__}",
         )
 
     def _fetch_session_obj(self) -> Dict[Any, Any]: