+++ test/unit-test/test_cli_arg.py
@@ -19,14 +19,13 @@ import os
 import pytest
 from _test_helper_functions import _test_command
 
-import seedfarmer.commands._stack_commands as _sc
-import seedfarmer.mgmt.module_init as minit
-from seedfarmer.__main__ import PROJECT, apply, destroy, init
+from seedfarmer import config
+from seedfarmer.__main__ import apply, destroy, init
 from seedfarmer.__main__ import list as _list
 from seedfarmer.__main__ import remove, store
 
-# Override _stack_commands OPS_ROOT to reflect path of resource policy needed for some testing #
-_sc.OPS_ROOT = os.path.join(_sc.OPS_ROOT, "test/unit-test/mock_data")
+# Override OPS_ROOT to reflect path of resource policy needed for some testing #
+config.OPS_ROOT = os.path.join(config.OPS_ROOT, "test/unit-test/mock_data")
 
 _logger: logging.Logger = logging.getLogger(__name__)
 
@@ -38,7 +37,7 @@ _logger: logging.Logger = logging.getLogger(__name__)
 @pytest.mark.init
 def test_init_create_module():
     module_name = "test-module"
-    expected_module_path = os.path.join(minit.OPS_ROOT, "modules")
+    expected_module_path = os.path.join(config.OPS_ROOT, "modules")
 
     # Creates a new module
     _test_command(sub_command=init, options=["module", "-m", module_name], exit_code=0, return_result=False)
@@ -56,7 +55,7 @@ def test_init_create_module():
 def test_init_create_group_module():
     module_name = "test-module"
     group_name = "group"
-    expected_module_path = os.path.join(minit.OPS_ROOT, "modules", group_name, module_name)
+    expected_module_path = os.path.join(config.OPS_ROOT, "modules", group_name, module_name)
 
     # Creates a group and a module within the group
     _test_command(sub_command=init, options=["module", "-g", group_name, "-m", module_name], exit_code=0)
@@ -68,7 +67,7 @@ def test_init_create_group_module():
     )
     assert (
         result.exception.args[0]
-        == f"The module {module_name} already exists under {minit.OPS_ROOT}/modules/{group_name}."
+        == f"The module {module_name} already exists under {config.OPS_ROOT}/modules/{group_name}."
     )
 
     # Checks if a file from the project template was created within the new module
@@ -77,14 +76,16 @@ def test_init_create_group_module():
 
 @pytest.mark.init
 def test_init_create_project(tmp_path):
-    expected_project_path = os.path.join(minit.OPS_ROOT, PROJECT)
+    expected_project_path = os.path.join(config.OPS_ROOT, config.PROJECT)
 
     # Creates a new project
     _test_command(sub_command=init, options=["project"], exit_code=0, return_result=False)
 
     # Creates a project that already exist
     result = _test_command(sub_command=init, options=["project"], exit_code=1, return_result=True)
-    assert result.exception.args[0] == f'Error: "{os.path.join(minit.OPS_ROOT, PROJECT)}" directory already exists'
+    assert (
+        result.exception.args[0] == f'Error: "{os.path.join(config.OPS_ROOT, config.PROJECT)}" directory already exists'
+    )
 
     # Checks if file exists from the project template
     assert os.path.exists(os.path.join(expected_project_path, "seedfarmer.yaml"))
@@ -101,7 +102,7 @@ def test_apply_help():
         sub_command=apply,
         options=["--help"],
         exit_code=0,
-        expected_output=f"Apply a deployment manifest relative path for {PROJECT.upper()}",
+        expected_output="Apply manifests to a SeedFarmer managed deployment",
     )
 
 
@@ -587,7 +588,7 @@ def test_store_moduledata_missing_module_arg():
 @pytest.mark.destroy
 def test_destroy():
     _test_command(
-        sub_command=destroy, options=["--help"], exit_code=0, expected_output=f"Destroy {PROJECT.upper()} Deployment"
+        sub_command=destroy, options=["--help"], exit_code=0, expected_output="Destroy a SeedFarmer managed deployment"
     )